+++ src/main/java/techcourse/myblog/web/LoginController.java
@@ -1,10 +1,15 @@
 package techcourse.myblog.web;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.dto.UserDto;
+import techcourse.myblog.exception.FailedLoginException;
 import techcourse.myblog.service.LoginService;
 
 import javax.servlet.http.HttpSession;
@@ -41,5 +46,12 @@ public class LoginController {
             httpSession.removeAttribute("email");
         }
         return "redirect:/";
+    }
+
+    @ExceptionHandler(FailedLoginException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public String handleFailedLoginException(FailedLoginException e, Model model) {
+        model.addAttribute("errorMessage", e.getMessage());
+        return "login";
     }
 }
